requirements.txt
South==0.8.4
django-mptt==0.6.0
